feat: add household finance CRUD, CSV import/export, and account deletion
Adds income, expense, budget, and savings-goal endpoints on top of the
ownership-enforced repository base: recurring entries store a rule once and
have their monthly occurrences computed on demand (packages/analytics), so
"how much recurs into this month" is never duplicated as stored rows.
Budgets expose computed planned-vs-actual status (spent, remaining,
over-budget); savings goals compute current amount, progress, and a
projected completion date from their contribution ledger rather than a
single running total.

CSV import is a real preview-then-commit flow: preview validates a file
(missing fields, bad amounts/dates, unknown categories, duplicates against
existing rows) without writing anything; commit re-validates and writes
only the valid, non-duplicate rows, returning a per-row report either way.
CSV export sanitizes cells starting with =, +, -, or @ against formula
injection. JSON account export covers every owned entity; DELETE /api/users/
me lets a user delete their own account (fastapi-users only ships a
superuser-gated delete-by-id route) and cascades all of their data.

Fixes a real route-ordering bug caught by the account-deletion tests: the
custom self-delete route has to be registered before fastapi-users' own
users router, otherwise its superuser-gated DELETE /{id} greedily matches
"me" as a literal id first and returns 403 instead of ever reaching ours.

52 backend tests passing; verified end-to-end against the running Docker
container (demo household's seeded budget shows correct actual/remaining).

+++ apps/api/app/models/reference_data.py
@@ -14,6 +14,7 @@ from __future__ import annotations
 import enum
 import uuid
 from datetime import datetime
+from decimal import Decimal
 
 from sqlalchemy import Boolean, DateTime, Enum, ForeignKey, Integer, Numeric, String, Text
 from sqlalchemy.dialects.postgresql import UUID
@@ -85,7 +86,7 @@ class SalarySnapshot(Base, UUIDPrimaryKeyMixin):
     import_run_id: Mapped[uuid.UUID] = mapped_column(
         UUID(as_uuid=True), ForeignKey("import_run.id", ondelete="CASCADE"), nullable=False
     )
-    median_gross: Mapped[float] = mapped_column(Numeric(10, 2), nullable=False)
+    median_gross: Mapped[Decimal] = mapped_column(Numeric(10, 2), nullable=False)
     year: Mapped[int] = mapped_column(Integer, nullable=False)
 
 
@@ -98,8 +99,8 @@ class RentSnapshot(Base, UUIDPrimaryKeyMixin):
     import_run_id: Mapped[uuid.UUID] = mapped_column(
         UUID(as_uuid=True), ForeignKey("import_run.id", ondelete="CASCADE"), nullable=False
     )
-    sqm_cold: Mapped[float] = mapped_column(Numeric(6, 2), nullable=False)
-    avg_apartment_size: Mapped[float] = mapped_column(Numeric(6, 2), nullable=False)
+    sqm_cold: Mapped[Decimal] = mapped_column(Numeric(6, 2), nullable=False)
+    avg_apartment_size: Mapped[Decimal] = mapped_column(Numeric(6, 2), nullable=False)
     year: Mapped[int] = mapped_column(Integer, nullable=False)
 
 
@@ -112,9 +113,9 @@ class CostSnapshot(Base, UUIDPrimaryKeyMixin):
     import_run_id: Mapped[uuid.UUID] = mapped_column(
         UUID(as_uuid=True), ForeignKey("import_run.id", ondelete="CASCADE"), nullable=False
     )
-    groceries_month: Mapped[float] = mapped_column(Numeric(8, 2), nullable=False)
-    transport_month: Mapped[float] = mapped_column(Numeric(8, 2), nullable=False)
-    utilities_month: Mapped[float] = mapped_column(Numeric(8, 2), nullable=False)
+    groceries_month: Mapped[Decimal] = mapped_column(Numeric(8, 2), nullable=False)
+    transport_month: Mapped[Decimal] = mapped_column(Numeric(8, 2), nullable=False)
+    utilities_month: Mapped[Decimal] = mapped_column(Numeric(8, 2), nullable=False)
     year: Mapped[int] = mapped_column(Integer, nullable=False)
 
 
@@ -127,7 +128,7 @@ class InflationSnapshot(Base, UUIDPrimaryKeyMixin):
     import_run_id: Mapped[uuid.UUID] = mapped_column(
         UUID(as_uuid=True), ForeignKey("import_run.id", ondelete="CASCADE"), nullable=False
     )
-    rate_pct: Mapped[float] = mapped_column(Numeric(5, 2), nullable=False)
+    rate_pct: Mapped[Decimal] = mapped_column(Numeric(5, 2), nullable=False)
     year: Mapped[int] = mapped_column(Integer, nullable=False)
